requirements.txt
spotipy>=2.24.0
python-dotenv>=1.0.0
rich>=13.0.0
